requirements.txt
discord.py>=2.3.0
gTTS>=2.3.0
python-dotenv>=1.0.0
PyNaCl>=1.5.0
google-cloud-texttospeech>=2.14.0
